requirements.txt
Flask
requests
urllib3
cryptography
obs-websocket-py
pywin32
psutil